pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["enekoegiguren <116632843+enekoegiguren@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"
plotly = "^5.24.1"
pillow = "^10.0.1"
boto3 = "^1.35.54"
seaborn = "^0.13.2"
scikit-learn = "^1.5.2"
joblib = "^1.4.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
